+++ src/libranet_logging/__init__.py
@@ -1,7 +1,9 @@
 """libranet_logging.__init__"""
 
 __version__ = "1.4.dev0"
+__copyright__ = "Copyright 2015 - 2023 Libranet"
+__license__ = "MIT License"
 
-from .cli import print_loggers, print_tree  # noqa
 from .logconfig import initialize, output_logging_tree  # noqa
 from .loglevel import create_loglevel, create_loglevel_trace  # noqa
+from .utils import print_loggers, print_tree  # noqa