+++ src/libranet_logging/yaml.py
@@ -72,7 +72,7 @@ def read_yaml(path, variables=None):
 
     # yaml.add_constructor("!env", constructor_env, yaml.SafeLoader)
 
-    with open(path, "r") as stream:
+    with open(path, "r", encoding="utf-8") as stream:
         try:
             data = stream.read()
             if variables: